Refactor evaluators into separate package
