+++ core/src/main/scala/fortress/operations/Simplification/DefinitionJit.scala
@@ -1,0 +1,10 @@
+package fortress.operations
+
+import fortress.msfol._
+
+/** Compiles function definitions to Scala functions on demand. */
+class DefinitionJit(theory: Theory) {
+
+
+
+}
