+++ core/src/main/scala/fortress/operations/Simplification/EvaluationInliner.scala
