+++ core/src/main/scala/fortress/operations/Simplification/Evaluator.scala
